refactor(core): promote pkg/completion to internal/core/completion (no behavior change)
