+++ cmd/databricks-claude/main.go
@@ -16,11 +16,11 @@ import (
 
 	"github.com/IceRhymers/databricks-agents/internal/cmd"
 	"github.com/IceRhymers/databricks-agents/internal/core/cli"
+	"github.com/IceRhymers/databricks-agents/internal/core/completion"
 	"github.com/IceRhymers/databricks-agents/internal/core/portbind"
 	"github.com/IceRhymers/databricks-agents/internal/core/refcount"
 	"github.com/IceRhymers/databricks-agents/internal/core/updater"
 	"github.com/IceRhymers/databricks-agents/pkg/authcheck"
-	"github.com/IceRhymers/databricks-agents/pkg/completion"
 	"github.com/IceRhymers/databricks-agents/pkg/health"
 	"github.com/IceRhymers/databricks-agents/pkg/mdmprofile"
 	"github.com/IceRhymers/databricks-agents/pkg/proxy"
